chat-lite-backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
boto3==1.40.11
botocore==1.40.11
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.99.9
packaging==25.0
passlib==1.7.4
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.7.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
s3transfer==0.13.1
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
